+++ pkarr/src/lib.rs
@@ -27,6 +27,7 @@ pub const DEFAULT_RELAYS: [&str; 2] = ["https://relay.pkarr.org", "https://pkarr
 /// Default [resolver](https://pkarr.org/resolvers)s
 pub const DEFAULT_RESOLVERS: [&str; 2] = ["resolver.pkarr.org:6881", "pkarr.pubky.app:6881"];
 
+#[cfg(any(target_arch = "wasm32", feature = "dht"))]
 pub use client::{Client, ClientBuilder, Settings};
 
 // Rexports