+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to pkarr client and server will be documented in this file.
 ### Added
 
 - Add strict monotonic unix `Timestamp`.
+- Impl `PartialEq, Eq` for `SignedPacket`.
+- Impl `From<PublicKey>` for `CacheKey`.
 
 ### Changed
 
@@ -16,6 +18,7 @@ All notable changes to pkarr client and server will be documented in this file.
 - replace `ureq` with `reqwest` to work with HTTP/2 relays, and Wasm.
 - update `mainline` to v3.0.0
 - `Client::shutdown` and `Client::shutdown_sync` are now idempotent and return `()`.
+- `Client::resolve`, `Client::resolve_sync` and `relay::Client::resolve` return expired cached `SignedPacket` if nothing else was found from the network.
 
 ### Removed